refactor(settings): split Metron and CBL panels
